refactor(browser): address PR review comments
