+++ packages/browser/package.json
@@ -16,10 +16,10 @@
   },
   "scripts": {
     "build": "tsup",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "format:check": "prettier --check \"src/**/*.ts\"",
+    "test": "cd ../.. && vitest run --project=unit tests/unit/browser.test.ts",
+    "test:watch": "cd ../.. && vitest --project=unit tests/unit/browser.test.ts",
+    "format": "prettier --write \"src/**/*.ts\" \"../../tests/unit/browser.test.ts\"",
+    "format:check": "prettier --check \"src/**/*.ts\" \"../../tests/unit/browser.test.ts\"",
     "clean": "rm -rf dist"
   },
   "author": "Langfuse",
@@ -31,6 +31,9 @@
   },
   "dependencies": {
     "@langfuse/core": "workspace:^"
+  },
+  "peerDependencies": {
+    "@opentelemetry/api": "^1.9.0"
   },
   "files": [
     "dist"
