--- packages/browser/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    environment: "happy-dom",
-    include: ["src/**/*.test.ts"],
-  },
-});
